+++ clickhouse-data/src/main/java/com/clickhouse/data/value/ClickHouseByteValue.java
@@ -85,7 +85,14 @@ public class ClickHouseByteValue implements ClickHouseValue {
 
         @Override
         public ClickHouseByteValue update(String value) {
-            return value == null ? resetToNullOrEmpty() : set(false, (byte) UnsignedByte.valueOf(value).byteValue());
+            if (value == null) {
+                resetToNullOrEmpty();
+            } else if (value.isEmpty()) {
+                resetToDefault();
+            } else {
+                set(false, UnsignedByte.valueOf(value).byteValue());
+            }
+            return this;
         }
     }
 
@@ -375,7 +382,14 @@ public class ClickHouseByteValue implements ClickHouseValue {
 
     @Override
     public ClickHouseByteValue update(String value) {
-        return value == null ? resetToNullOrEmpty() : set(false, Byte.parseByte(value));
+        if (value == null) {
+            resetToNullOrEmpty();
+        } else if (value.isEmpty()) {
+            resetToDefault();
+        } else {
+            set(false, Byte.parseByte(value));
+        }
+        return this;
     }
 
     @Override