+++ clickhouse-data/src/main/java/com/clickhouse/data/value/ClickHouseFloatValue.java
@@ -274,7 +274,14 @@ public class ClickHouseFloatValue implements ClickHouseValue {
 
     @Override
     public ClickHouseFloatValue update(String value) {
-        return value == null ? resetToNullOrEmpty() : set(false, Float.parseFloat(value));
+        if (value == null) {
+            resetToNullOrEmpty();
+        } else if (value.isEmpty()) {
+            resetToDefault();
+        } else {
+            set(false, Float.parseFloat(value));
+        }
+        return this;
     }
 
     @Override