+++ clickhouse-data/src/main/java/com/clickhouse/data/value/ClickHouseMapValue.java
@@ -140,7 +140,7 @@ public class ClickHouseMapValue extends ClickHouseObjectValue<Map<?, ?>> {
     public String asString() {
         Map<?, ?> value = getValue();
         if (value == null || value.isEmpty()) {
-            return "{}";
+            return ClickHouseValues.EMPTY_MAP_EXPR;
         }
         StringBuilder builder = new StringBuilder().append('{');
         for (Entry<?, ?> e : value.entrySet()) {
@@ -171,7 +171,7 @@ public class ClickHouseMapValue extends ClickHouseObjectValue<Map<?, ?>> {
     public String toSqlExpression() {
         Map<?, ?> value = getValue();
         if (value == null || value.isEmpty()) {
-            return "{}";
+            return ClickHouseValues.EMPTY_MAP_EXPR;
         }
 
         StringBuilder builder = new StringBuilder().append('{');
@@ -248,10 +248,15 @@ public class ClickHouseMapValue extends ClickHouseObjectValue<Map<?, ?>> {
     @Override
     public ClickHouseMapValue update(String value) {
         if (value == null) {
-            return resetToNullOrEmpty();
+            resetToNullOrEmpty();
+        } else if (value.isEmpty() || ClickHouseValues.EMPTY_MAP_EXPR.equals(value)) {
+            resetToDefault();
+        } else {
+            // TODO parse string
+            set(Collections.singletonMap(getDefaultKey(), valueType.cast(value)));
         }
 
-        return set(Collections.singletonMap(getDefaultKey(), valueType.cast(value)));
+        return this;
     }
 
     @Override