+++ clickhouse-data/src/main/java/com/clickhouse/data/value/ClickHouseNestedValue.java
@@ -532,10 +532,14 @@ public class ClickHouseNestedValue extends ClickHouseObjectValue<Object[][]> {
     @Override
     public ClickHouseNestedValue update(String value) {
         if (value == null) {
-            return resetToNullOrEmpty();
+            resetToNullOrEmpty();
+        } else if (value.isEmpty() || ClickHouseValues.EMPTY_ARRAY_EXPR.equals(value)) {
+            resetToDefault();
+        } else {
+            // TODO parse string
+            set(new Object[][] { new String[] { value } });
         }
-        // TODO parse string
-        return set(new Object[][] { new String[] { value } });
+        return this;
     }
 
     @Override